Add an optional [properties] in waydroid.cfg
On some platforms, like a VM, we don't have the "getprop" android
utility available. It could be, also, that the default values
for some properties don't match the target device.

With this PR, one can add a [properties] section in waydroid.cfg,
that will act as an override for the values that would have been
put otherwise.

Example, on our LuneOS x86 VM:
[properties]
ro.hardware.gralloc=default
ro.hardware.egl=mesa

A minimal waydroid.cfg could even be shipped with waydroid, and
the rest of the configuration will be filled with "waydroid init".

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>