Support upgrading major android versions
When restarting the lxc container our binder services will receive a
binder_presence event and will try to re-register on the service manager.
If the service manager was created with aidl2 and the new android version
uses aidl3 this will fail with:
  Failed to add service waydroidusermonitor: -2147483647

When that happens, try to restart the service by getting a new service
manager that can handle the new aidl protocol version.