Mount overlays on top of image mount points
This allows the user to make modifications to the images that may persist
between image upgrades.

For both the system and vendor image there's a set of two overlays.
One, specified in config as "overlay", is a read-only persistent overlay
meant for stuff like installing privileged apps that should persist.
Second one, specified as "overlay-rw", is a read-write overlay that
stores the changes made by the user in case they remount the mount point
in read-write mode. This one is meant to be removed when performing
image upgrade to not carry on potentially incompatible changes between
images.