systemd: wait for org.gnome.SessionManager on session bus
we start waydroid-session service as furios if autostart is enabled. early phosh start reports not 100% accurate numbers to xdg toplevel because phosh is creating the status and navbar so its a few pixels too big when android asks for the value. wait for org.gnome.SessionManager (which is provided by phosh itself) and wait a bit more after that to make sure we get a good value

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>