statechange_server: implement composer monitor and kill composer if it gets stuck on startup
this is a workaround and should be dropped when we figure out why composer does not register onto binder properly at times.
check if the composer extension interface is up, if it is not then kill composer and let it start again. this has been proven to work but it is a very dirty hack.

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>